pyproject.toml
[project]
name = "oc-p6-rugby-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=1.1.7",
    "imbalanced-learn>=0.14.0",
    "ipykernel>=7.1.0",
    "jupyterlab>=4.4.10",
    "matplotlib>=3.10.0",
    "mlflow>=3.5.1",
    "numpy>=2.1.3",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "tqdm>=4.67.1",
    "xgboost>=3.1.1",
    "ydata-profiling>=4.17.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "pre-commit>=4.3.0",
    "ruff>=0.14.3",
]
